Inline CMIP6 base classes into plugin model/record; remove plugins/cmip6/base; add lazy re-exports to avoid circular imports; tests+lint green
